Add support for static linters
